requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
twilio==8.10.0
phonenumbers==8.13.21
requests==2.31.0
pytz==2024.1
openai==1.14.3
